fix portability issues on mobile by requesting a highp precision!

+++ src/render/src/shaders/grid/mercator.frag
@@ -1,5 +1,5 @@
 #version 300 es
-precision lowp float;
+precision highp float;
 
 out vec4 c;
 in vec2 pos_clip;
